+++ frontend/src/pages/feedback/FeedbackEdit.tsx
@@ -4,10 +4,10 @@ import useFeedbackEdit from 'hooks/feedback/useFeedbackEdit';
 import useRole from 'hooks/useRole';
 import useContentTag from 'hooks/utils/useContentTag';
 
-import ContentHeader from 'components/@commons/ContentHeader';
 import FlexBox from 'components/@commons/FlexBox';
 import ToolTip from 'components/@commons/ToolTip';
 import BottomBar from 'components/@commons/bottomBar/BottomBar';
+import ContentHeader from 'components/@commons/contentHeader/ContentHeader';
 import WriterDocument from 'components/WriterDocument';
 import FeedbackFormat from 'components/feedbacks/FeedbackFormat';
 import InterviewQuestion from 'components/interviewQuestion/InterviewQuestion';
