+++ frontend/src/pages/feedback/Feedbacks.tsx
@@ -15,9 +15,9 @@ import Loading from 'pages/status/Loading';
 import plusIcon from 'assets/images/plus.svg';
 import { TEAM_STATUS } from 'constants/constants';
 
-import ContentHeader from 'components/@commons/ContentHeader';
 import Image from 'components/@commons/Image';
 import Button from 'components/@commons/button/Button';
+import ContentHeader from 'components/@commons/contentHeader/ContentHeader';
 import Feedback from 'components/feedbacks/Feedback';
 import { FeedbackType } from 'types/feedback';
 import { feedbackAddUriBuilder, teamGetUriBuilder } from 'utils/util';
