+++ frontend/src/pages/interviewQuestion/InterviewQuestions.tsx
@@ -13,8 +13,8 @@ import Loading from 'pages/status/Loading';
 import { TEAM_STATUS } from 'constants/constants';
 import { GITHUB_AVATAR_SIZE_LIST } from 'constants/constants';
 
-import ContentHeader from 'components/@commons/ContentHeader';
 import Image from 'components/@commons/Image';
+import ContentHeader from 'components/@commons/contentHeader/ContentHeader';
 import {
   InterviewQuestionsInLevellogType,
   InterviewQuestionInfoType,
